model seperation and api
I decided to seperate the machine learning model to another django project, the version will be checked and the latest loaded model will be updated via api. I added views to request for the version and download if there is a newer one available. i also moved all the model data analysis scripts, analysis results and loaded models to the new project which i will upload to github